code/requirements.txt
seqeval==1.2.2
torch==2.9.1
torchtext==0.6.0
tensorflow==2.20.0
keras==3.12.0
pandas==2.3.3
transformers==4.57.3
click==8.3.1
nltk==3.9.2
sklearn-crfsuite==0.5.0
catboost==1.2.8
lightgbm==4.6.0
python-telegram-bot==22.5